pyproject.toml
[project]
name = "semantic-search-engine"
version = "0.1.0"
description = "Semantic PDF search engine using transformer embeddings and FAISS."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Amirali Ghafarian" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "python-dotenv>=1.0.1",
  "pymupdf>=1.24.9",
  "sentence-transformers>=3.0.1",
  "faiss-cpu>=1.8.0.post1",
  "numpy>=1.26.4",
  "rank-bm25>=0.2.2"
]

[dependency-groups]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.2",
  "mypy>=1.11.1",
  "httpx>=0.27.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
ignore_missing_imports = true
strict = false